% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_backtest}
\alias{cpp_backtest}
\title{Fast Backtest Engine in C++}
\usage{
cpp_backtest(asset_returns, signals)
}
\arguments{
\item{asset_returns}{A numeric vector of the daily percentage returns of the asset.}

\item{signals}{A numeric vector of signals (-1, 0, 1). Signals should theoretically be lagged by 1 day to prevent look-ahead bias.}
}
\value{
A numeric vector of the cumulative portfolio equity curve.
}
\description{
Iterates through a vector of returns and trading signals to compute the cumulative equity line.
}
