% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{fetch_asset_data}
\alias{fetch_asset_data}
\title{Fetch Financial Data}
\usage{
fetch_asset_data(symbol, start_date = "2020-01-01", end_date = Sys.Date())
}
\arguments{
\item{symbol}{A character string representing the ticker symbol (e.g., "SPY").}

\item{start_date}{The starting date for the data (e.g., "2020-01-01").}

\item{end_date}{The ending date for the data (e.g., Sys.Date()).}
}
\value{
A tidy dataframe (tibble) of OHLCV daily data.
}
\description{
Leverages tidyquant to download historical price data for stocks/commodities.
We use yahoo finance API primarily for North American trading (TSX and US).
}
