% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_strategy_discovery.R
\name{mod_strategy_discovery_ui}
\alias{mod_strategy_discovery_ui}
\title{Strategy Discovery UI}
\usage{
mod_strategy_discovery_ui(id)
}
\arguments{
\item{id}{NA}
}
\description{
Strategy Discovery UI
}
